patient & visit changes

+++ GSC.Data/Entities/Master/PatientStatus.cs
@@ -1,11 +1,12 @@
 ﻿using GSC.Data.Entities.Common;
+using System;
 
 namespace GSC.Data.Entities.Master
 {
     public class PatientStatus : BaseEntity
     {
 
-        public int Code { get; set; }
+        public Int16 Code { get; set; }
         public string StatusName { get; set; }
         public int? CompanyId { get; set; }
     }
